Task 38 - Refactoring game module

+++ src/main/java/pl/stormit/eduquiz/game/service/GameService.java
@@ -3,7 +3,7 @@ package pl.stormit.eduquiz.game.service;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import org.springframework.web.client.ResourceAccessException;
+import org.springframework.transaction.annotation.Transactional;
 import pl.stormit.eduquiz.game.domain.entity.Game;
 import pl.stormit.eduquiz.game.domain.repository.GameRepository;
 import pl.stormit.eduquiz.game.dto.GameDto;
@@ -20,6 +20,7 @@ import java.util.UUID;
 @Service
 @RequiredArgsConstructor
 public class GameService {
+
     private final QuizRepository quizRepository;
 
     private final GameRepository gameRepository;
@@ -30,19 +31,21 @@ public class GameService {
 
     List<UUID> userAnswers = new ArrayList<>();
 
-    public GameDto createGame(QuizDto quizRequest) {
-        Quiz quiz = quizRepository.findById(quizRequest.id()).orElseThrow(() -> {
-            throw new RuntimeException("The quiz does not exist");
+    @Transactional
+    public GameDto createGame(QuizDto quizRequestDto) {
+        Quiz quiz = quizRepository.findById(quizRequestDto.id()).orElseThrow(() -> {
+            throw new EntityNotFoundException("The quiz by id: " + quizRequestDto.id() + ", does not exist.");
         });
         Game game = new Game(quiz);
         game.setUserAnswers(userAnswers);
 
-        return gameMapper.mapGameEntityToGameDto(gameRepository.save(game));
+        return gameMapper.mapGameEntityToGameDto(game);
     }
 
+    @Transactional(readOnly = true)
     public GameDto getGame(UUID gameId) {
         Game game = gameRepository.findById(gameId).orElseThrow(() -> {
-            throw new EntityNotFoundException("The game does not exist");
+            throw new EntityNotFoundException("The game by id: " + gameId + ", does not exist.");
         });
         return gameMapper.mapGameEntityToGameDto(game);
     }
@@ -50,7 +53,7 @@ public class GameService {
     public GameDto playGame(UUID gameId, AnswerDto answer) {
         playGame = gameRepository.findById(gameId)
                 .orElseThrow(() -> {
-                    throw new ResourceAccessException("The quiz does not exist");
+                    throw new EntityNotFoundException("The game by id: " + gameId + ", does not exist.");
                 });
 
         userAnswers.add(answer.id());
@@ -59,17 +62,23 @@ public class GameService {
         return gameMapper.mapGameEntityToGameDto(playGame);
     }
 
+    @Transactional
     public GameDto completeGame(UUID gameId) {
         Game updateGame = gameRepository.findById(gameId)
                 .orElseThrow(() -> {
-                    throw new ResourceAccessException("The game does not exist");
+                    throw new EntityNotFoundException("The game by id: " + gameId + ", does not exist.");
                 });
         updateGame.setUserAnswers(playGame.getUserAnswers());
 
-        return gameMapper.mapGameEntityToGameDto(gameRepository.save(updateGame));
+        return gameMapper.mapGameEntityToGameDto(updateGame);
     }
 
+    @Transactional
     public void deleteGame(UUID gameId) {
-        gameRepository.deleteById(gameId);
+        if (gameRepository.existsById(gameId)) {
+            gameRepository.deleteById(gameId);
+        } else {
+            throw new EntityNotFoundException("The game by id: " + gameId + ", does not exist.");
+        }
     }
 }
